Updated RecordParser so it can be converted to other languages by codegen
